fix: separete commands from root
