+++ packages/aspida/src/commands.ts
@@ -4,6 +4,9 @@ import buildTemplate from './buildTemplate'
 import writeRouteFile from './writeRouteFile'
 import watchInputDir from './watchInputDir'
 
+export { getConfigs, AspidaConfig } from './getConfigs'
+export { parse } from './parseInterface'
+
 export const version = () => require(path.join(__dirname, '../../package.json')).version
 
 export const build = (config?: Parameters<typeof getConfigs>[0]) =>
